Ensure there is always a default for date timezone (EC2)
Sometimes on an Amazon EC2 server the date default timezone is not set,
which can cause date-related operations to throw an exception. We want
to silently fail in this case and load a default timezone from
configuration; if this is not specified, we fall back to calling
date_default_timezone_get() which would return "UTC" if the timezone is
not set.

We only need this single line because if there is a default timezone
set, date_default_timezone_get() returns that.