Ensure that (empty?) content is rendered for tags that must have content
Certain browsers (or a braindead W3C specification) do not allow certain
tags to be declared empty. Off the top of my head I can name the SCRIPT
tag, the DIV tag, and the I tag (in Firefox 48.0 at least). If there are
any more, this is where they should be added so Phast knows to ALWAYS
render them with content, even when content is empty.