Make the suggestion list automatically expand to the largest element
WARNING: Due to the -moz- prefix, it can be assumed that this "solution"
only works with Mozilla/Gecko-based browsers. We need to find a
permanent cross-browser solution for this, hopefully one that does not
involve hacking with JavaScript.

P.S. first of all, why is this container NOT automatically resizing to
fit its children? Aren't they supposed to do that automatically? Is
there some style overriding this that I am overlooking?